+++ src/sumo/modes/run/solvers/__init__.py
@@ -1,0 +1,5 @@
+from .unsupervised_sumo import UnsupervisedSumoNMF
+from .supervised_sumo import SupervisedSumoNMF
+
+__all__ = ["UnsupervisedSumoNMF",
+           "SupervisedSumoNMF"]